// app/enums/meta-tag.enum.ts
export enum MetaProperty {
  OgTitle = "og:title",
  OgDescription = "og:description",
  OgType = "og:type",
  OgUrl = "og:url",
  OgImage = "og:image",
  OgImageAlt = "og:image:alt",
  OgSiteName = "og:site_name",
}

export enum MetaName {
  Description = "description",
  TwitterCard = "twitter:card",
  TwitterTitle = "twitter:title",
  TwitterDescription = "twitter:description",
  TwitterImage = "twitter:image",
}

export enum TwitterCardType {
  Summary = "summary",
  SummaryLargeImage = "summary_large_image",
}